+++ ext/ClimaAnalysisGeoMakieExt.jl
@@ -7,17 +7,42 @@ import ClimaAnalysis: Visualize
 
 MakiePlace = Union{Makie.Figure, Makie.GridLayout}
 
+"""
+    oceanmask()
+
+Return a collection of polygons to mask out the ocean.
+
+Plot with `Makie.poly`.
+"""
+function Visualize.oceanmask()
+    elevation = 0
+    return GeoMakie.NaturalEarth.bathymetry(elevation)
+end
+
+"""
+    landmask()
+
+Return a collection of polygons to mask out the continents.
+
+Plot with `Makie.poly`.
+"""
+function Visualize.landmask()
+    return GeoMakie.land()
+end
+
 function _geomakie_plot_on_globe!(
     place::MakiePlace,
     var::ClimaAnalysis.OutputVar;
     p_loc = (1, 1),
     plot_coastline = true,
     plot_colorbar = true,
+    mask = nothing,
     more_kwargs = Dict(
         :plot => Dict(),
         :cb => Dict(),
         :axis => Dict(),
         :coast => Dict(:color => :black),
+        :mask => Dict(),
     ),
     plot_fn = Makie.surface!,
 )
@@ -47,6 +72,9 @@ function _geomakie_plot_on_globe!(
     plot_kwargs = get(more_kwargs, :plot, Dict())
     cb_kwargs = get(more_kwargs, :cb, Dict())
     coast_kwargs = get(more_kwargs, :coast, Dict(:color => :black))
+    mask_kwargs = get(more_kwargs, :mask, Dict(:color => :white))
+
+    plot_mask = !isnothing(mask)
 
     var.attributes["long_name"] =
         ClimaAnalysis.Utils.warp_string(var.attributes["long_name"])
@@ -56,6 +84,7 @@ function _geomakie_plot_on_globe!(
     GeoMakie.GeoAxis(place[p_loc...]; title, axis_kwargs...)
 
     plot = plot_fn(lon, lat, var.data; plot_kwargs...)
+    plot_mask && Makie.poly!(mask; mask_kwargs...)
     plot_coastline && Makie.lines!(GeoMakie.coastlines(); coast_kwargs...)
 
     if plot_colorbar
@@ -75,12 +104,14 @@ end
                         p_loc = (1,1),
                         plot_coastline = true,
                         plot_colorbar = true,
+                        mask = nothing,
                         more_kwargs)
     heatmap2D_on_globe!(grid_layout::Makie.GridLayout,
                         var::ClimaAnalysis.OutputVar;
                         p_loc = (1,1),
                         plot_coastline = true,
                         plot_colorbar = true,
+                        mask = nothing,
                         more_kwargs)
 
 
@@ -95,6 +126,13 @@ This function assumes that the following attributes are available:
 
 The dimensions have to be longitude and latitude.
 
+`mask` has to be an object that can be plotted by `Makie.poly`. Typically, an ocean or land
+mask. `ClimaAnalysis` comes with predefined masks, check out [`Visualize.oceanmask`](@ref) and
+[`Visualize.landmask`](@ref).
+
+!!! note Masking does not affect the colorbar. If you have values defined beneath the map,
+    they can still affect the colorbar.
+
 Additional arguments to the plotting and axis functions
 =======================================================
 
@@ -103,6 +141,7 @@ Additional arguments to the plotting and axis functions
 - the plotting function (`:plot`)
 - the colorbar (`:cb`)
 - the coastline (`:coast`)
+- the mask (`:mask`)
 
 The coastline is plotted from `GeoMakie.coastline` using the `lines!` plotting function.
 
@@ -115,11 +154,13 @@ function Visualize.heatmap2D_on_globe!(
     p_loc = (1, 1),
     plot_coastline = true,
     plot_colorbar = true,
+    mask = nothing,
     more_kwargs = Dict(
         :plot => Dict(),
         :cb => Dict(),
         :axis => Dict(),
         :coast => Dict(:color => :black),
+        :mask => Dict(),
     ),
 )
     return _geomakie_plot_on_globe!(
@@ -128,6 +169,7 @@ function Visualize.heatmap2D_on_globe!(
         p_loc,
         plot_coastline,
         plot_colorbar,
+        mask,
         more_kwargs,
         plot_fn = Makie.surface!,
     )
@@ -140,6 +182,7 @@ end
                         plot_coastline = true,
                         plot_colorbar = true,
                         plot_contours = true,
+                        mask = nothing,
                         more_kwargs)
     contours2D_on_globe!(grid_layout::Makie.GridLayout,
                         var::ClimaAnalysis.OutputVar;
@@ -147,6 +190,7 @@ end
                         plot_coastline = true,
                         plot_colorbar = true,
                         plot_contours = true,
+                        mask = nothing,
                         more_kwargs)
 
 
@@ -161,6 +205,13 @@ This function assumes that the following attributes are available:
 
 The dimensions have to be longitude and latitude.
 
+`mask` has to be an object that can be plotted by `Makie.poly`. Typically, an ocean or land
+mask. `ClimaAnalysis` comes with predefined masks, check out [`Visualize.oceanmask`](@ref) and
+[`Visualize.landmask`](@ref).
+
+!!! note Masking does not affect the colorbar. If you have values defined beneath the map,
+    they can still affect the colorbar.
+
 Additional arguments to the plotting and axis functions
 =======================================================
 
@@ -169,6 +220,7 @@ Additional arguments to the plotting and axis functions
 - the plotting function (`:plot`)
 - the colorbar (`:cb`)
 - the coastline (`:coast`)
+- the mask (`:mask`)
 
 The coastline is plotted from `GeoMakie.coastline` using the `lines!` plotting function.
 
@@ -181,11 +233,13 @@ function Visualize.contour2D_on_globe!(
     p_loc = (1, 1),
     plot_coastline = true,
     plot_colorbar = true,
+    mask = nothing,
     more_kwargs = Dict(
         :plot => Dict(),
         :cb => Dict(),
         :axis => Dict(),
         :coast => Dict(:color => :black),
+        :mask => Dict(),
     ),
 )
     _geomakie_plot_on_globe!(
@@ -194,6 +248,7 @@ function Visualize.contour2D_on_globe!(
         p_loc,
         plot_coastline,
         plot_colorbar,
+        mask,
         more_kwargs,
         plot_fn = Makie.contourf!,
     )